확인후 업로드 기능 추가

+++ uploadCardnewsToInstagram.py
@@ -3,40 +3,45 @@ import glob
 from dotenv import load_dotenv
 from instabot import Bot
 
-cookie_del = glob.glob("config/*cookie.json")
-os.remove(cookie_del[0])
+from config import HASHTAG
 
-# 폴더 이름 입력 받기
-folder_name = input("업로드할 폴더의 이름을 입력하세요: ")  # 폴더 이름을 입력받습니다.
+def uploadCardnewsToInstagram(folder_name):
+    # Delete existing cookie files
+    cookie_del = glob.glob("config/*cookie.json")
+    if cookie_del:
+        os.remove(cookie_del[0])
 
-# load .env
-load_dotenv()
+    # Load .env
+    load_dotenv()
+    username = os.environ.get('YOUR_INSTA_USERNAME')
+    password = os.environ.get('YOUR_INSTA_PASSWORD')
 
-username = os.environ.get('YOUR_INSTA_USERNAME')
-password = os.environ.get('YOUR_INSTA_PASSWORD')
+    # Create an Instabot instance
+    bot = Bot()
 
-# Instabot 객체 생성
-bot = Bot()
+    # Instagram account login
+    bot.login(username=username, password=password)
 
-# Instagram 계정 로그인 (자신의 Instagram 계정 아이디와 비밀번호를 입력하세요)
-bot.login(username=username, password=password)
+    # Get a list of image files in the specified folder
+    photos = []
+    for filename in os.listdir(folder_name):
+        if filename.endswith(('.jpg', '.jpeg', '.png')):
+            photos.append(os.path.join(folder_name, filename))
 
-# 폴더 내의 모든 이미지 파일 가져오기
-photos = []
-for filename in os.listdir(folder_name):
-    if filename.endswith(('.jpg', '.jpeg', '.png')):  # 이미지 파일 확장자를 확인
-        photos.append(os.path.join(folder_name, filename))
+    # Set the caption based on the folder name and hashtag
+    caption_title = f"{folder_name}"
+    hashtag = HASHTAG
+    caption = caption_title + hashtag
 
-# 폴더 이름을 사용하여 caption_title 설정
-caption_title = f"{folder_name}"  # 폴더 이름을 제목에 사용합니다.
-hashtag = "  . . .     #ENFP#ISTJ#ISFJ#INFJ#INTJ#ISTP#ISFP#INFP#INTP#ESTP#ESFP#ENTP#ESTJ#ESFJ#ENFJ#ENTJ#인스타툰"
+    # Upload photos
+    for photo in photos:
+        bot.upload_photo(photo, caption=caption)
 
-caption = caption_title + hashtag
+    # Remove temporary files after upload (optional)
+    for photo in photos:
+        os.remove(photo + ".REMOVE_ME")
 
-# 사진 업로드
-for photo in photos:
-    bot.upload_photo(photo, caption=caption)
-
-# 업로드 후 임시 파일 제거 (선택 사항)
-for photo in photos:
-    os.remove(photo + ".REMOVE_ME")
+# Test code
+if __name__ == "__main__":
+    folder_name = input("Enter the name of the folder to upload: ")
+    uploadCardnewsToInstagram(folder_name)
